+++ tck/src/main/java/ee/jakarta/tck/concurrent/spec/ContextService/contextPropagate/servlet/BaseTestRunnableWork.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package ee.jakarta.tck.concurrent.spec.ContextService.contextPropagate_servlet;
+package ee.jakarta.tck.concurrent.spec.ContextService.contextPropagate.servlet;
 
 import java.io.Serializable;
