+++ tck/src/main/java/ee/jakarta/tck/concurrent/spec/ContextService/contextPropagate/servlet/TestJNDIRunnableWork.java
@@ -14,12 +14,20 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package ee.jakarta.tck.concurrent.spec.ContextService.contextPropagate_servlet;
+package ee.jakarta.tck.concurrent.spec.ContextService.contextPropagate.servlet;
 
-import java.io.Serializable;
+import javax.naming.InitialContext;
+import javax.naming.NamingException;
 
-public interface TestWorkInterface extends Runnable, Serializable {
-	public String doSomeWork();
+@SuppressWarnings("serial")
+public class TestJNDIRunnableWork extends BaseTestRunnableWork {
 
-	public String getResult();
+	@Override
+	protected String work() {
+		try {
+            return InitialContext.doLookup("java:comp/env/myMessage");
+        } catch (NamingException e) {
+            throw new RuntimeException(e);
+        }
+	}
 }
