+++ tck/src/main/java/ee/jakarta/tck/concurrent/spec/ManagedThreadFactory/context/servlet/ContextServletTests.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package ee.jakarta.tck.concurrent.spec.ManagedThreadFactory.context_servlet;
+package ee.jakarta.tck.concurrent.spec.ManagedThreadFactory.context.servlet;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
